added markerpolicy

+++ site/app/Http/Controllers/MarkerController.php
@@ -5,13 +5,13 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Requests\MarkerRequest;
 use App\Marker;
-
+use Auth;
 
 class MarkerController extends Controller
 {
     public function index()
     {
-    	$markers = Marker::all();
+    	$markers = Auth::user()->markers;
 
     	return view('marker.index', compact('markers'));
     }
@@ -19,13 +19,22 @@ class MarkerController extends Controller
 
     public function create()
     {
+    	if (!Auth::user()->can('create', Marker::class)) {
+            return abort(403);
+        }
+
         return view('marker.create');
     }
 
 
     public function store(MarkerRequest $request)
     {
+    	if (!Auth::user()->can('store', Marker::class)) {
+            return abort(403);
+        }
+
     	$marker = new Marker;
+    	$marker->user_id = Auth::id();
     	$marker->latitude = $request->latitude;
     	$marker->longitude = $request->longitude;
     	$marker->name = $request->name;
@@ -42,6 +51,10 @@ class MarkerController extends Controller
     {
     	$marker = Marker::findOrFail($id);
 
+    	if (!Auth::user()->can('view', $marker)) {
+            return abort(403);
+        }
+
         return view('marker.show', compact('marker'));
     }
 
@@ -50,6 +63,10 @@ class MarkerController extends Controller
     {
     	$marker = Marker::findOrFail($id);
 
+    	if (!Auth::user()->can('edit', $marker)) {
+            abort(403);
+        }
+
         return view('marker.edit', compact('marker'));
     }
 
@@ -57,6 +74,10 @@ class MarkerController extends Controller
     public function update(MarkerRequest $request, $id)
     {
     	$marker = Marker::findOrFail($id);
+
+    	if (!Auth::user()->can('update', $marker)) {
+            abort(403);
+        }
 
     	$marker->latitude = $request->latitude;
     	$marker->longitude = $request->longitude;
@@ -73,6 +94,10 @@ class MarkerController extends Controller
     public function destroy(Request $request, $id)
     {
     	$marker = Marker::findOrFail($id);
+
+    	if (!Auth::user()->can('destroy', $marker)) {
+            abort(403);
+        }
 
     	$marker->delete();
